+++ cli/src/commit_templater.rs
@@ -2131,6 +2131,30 @@ fn builtin_change_id_methods<'repo>() -> CommitTemplateBuildMethodFnMap<'repo, C
             Ok(out_property.into_dyn_wrapped())
         },
     );
+    map.insert(
+        "remote_bookmarks",
+        |language, _diagnostics, _build_ctx, self_property, function| {
+            function.expect_no_arguments()?;
+            let repo = language.repo;
+            let index = language
+                .keyword_cache
+                .bookmarks_index(language.repo)
+                .clone();
+            let out_property = self_property.and_then(move |change_id| {
+                let maybe_targets = repo.resolve_change_id(&change_id).block_on()?;
+                Ok(maybe_targets
+                    .iter()
+                    .flat_map(|targets| targets.targets.iter())
+                    .flat_map(|(commit_id, _state)| collect_remote_refs(index.get(commit_id)))
+                    // Since we are iterating through commits here, a confliciting remote bookmark
+                    // could point two different commits for the same change ID. Check here for
+                    // uniqueness so we don't return the same bookmark twice.
+                    .unique_by(Rc::as_ptr)
+                    .collect_vec())
+            });
+            Ok(out_property.into_dyn_wrapped())
+        },
+    );
     map
 }